Updated to correct a few errors & added some markdown.
I'm finding the README, the programs and the extra source code to be a bit out of sync. I've changed the start program section was a bit off (v09 defaults to v09.rom, I've added a -f option to handle that). I've added some more markdown and the Notes, License and update other sections.